code contrasts partial least squares regression (plsr) vs principal components regression (pcr) for predicting gender based on body measurements
PLSR and PCR are both methods to model a response variable when there
are a large number of predictor variables, and those predictors are
highly correlated or even collinear. Both methods construct new
predictor variables, known as components, as linear combinations of the
original predictor variables, but they construct those components in
different ways. PCR creates components to explain the observed
variability in the predictor variables, without considering the
response variable at all. PLSR does take the response variable into
account, and can lead to models that are able to fit the response
variable with fewer components.